Add navbar to switch between classic and new views
