+++ src/components/Animanga/Animanga3M.js
@@ -1,0 +1,8 @@
+export default function Animanga3M({id}) {
+
+    return (
+        <span>
+            XD
+        </span>
+    );
+}
