+++ src/components/User/User.css
@@ -4,7 +4,6 @@ header.user {
     background-repeat: no-repeat;
     background-size: cover;
     height: 240px;
-    margin-bottom: 12px;
     position: relative;
     display: flex;
     align-items: flex-end;
